ml_cli/api/main.py
from fastapi import FastAPI
from pydantic import BaseModel
import pandas as pd
from sklearn.pipeline import make_pipeline
from sklearn.preprocessing import StandardScaler
from sklearn.ensemble import RandomForestClassifier

app = FastAPI()

class PredictionInput(BaseModel):
    # Define the input schema for your model
    # This should match the columns of your training data
    # Example:
    feature1: float
    feature2: float

@app.post("/predict")
def predict(input_data: PredictionInput):
    """Make predictions using the trained model."""
    try:
        # Convert input data to a pandas DataFrame
        input_df = pd.DataFrame([input_data.dict()])

        # This is a placeholder for loading your actual trained model
        # In a real scenario, you would load the saved model here
        # For example, using joblib or pickle
        # from sklearn.externals import joblib
        # model = joblib.load('path/to/your/model.pkl')

        # For demonstration, we create and fit a dummy pipeline
        # This should be replaced by your actual trained pipeline
        pipeline = make_pipeline(StandardScaler(), RandomForestClassifier(random_state=42))
        
        # Fit the pipeline with dummy data to simulate a trained model
        from sklearn.datasets import make_classification
        X_dummy, y_dummy = make_classification(n_samples=100, n_features=len(input_df.columns), n_informative=len(input_df.columns), n_redundant=0, random_state=42)
        pipeline.fit(X_dummy, y_dummy)

        # Make predictions
        prediction = pipeline.predict(input_df)

        return {"prediction": prediction.tolist()}

    except Exception as e:
        return {"error": str(e)}

@app.get("/")
def read_root():
    return {"message": "Welcome to the ML Model API"}
